// projects/shared/src/lib/enums/medical-specialty.ts
export enum MedicalSpecialty {
    CARDIOLOGY = 'CARDIOLOGY',
    DERMATOLOGY = 'DERMATOLOGY',
    ENDOCRINOLOGY = 'ENDOCRINOLOGY',
    GASTROENTEROLOGY = 'GASTROENTEROLOGY',
    GERIATRICS = 'GERIATRICS',
    GYNECOLOGY = 'GYNECOLOGY',
    HEMATOLOGY = 'HEMATOLOGY',
    INFECTIOUS_DISEASES = 'INFECTIOUS_DISEASES',
    NEPHROLOGY = 'NEPHROLOGY',
    NEUROLOGY ='NEUROLOGY',
    OPHTHALMOLOGY = 'OPHTHALMOLOGY',
    ONCOLOGY = 'ONCOLOGY',
    ORTHOPEDICS ='ORTHOPEDICS',
    OTOLARYNGOLOGY = 'OTOLARYNGOLOGY',
    PEDIATRICS ='PEDIATRICS',
    PULMONOLOGY = 'PULMONOLOGY',
    PSYCHIATRY = 'PSYCHIATRY',
    PSYCHOLOGY = 'PSYCHOLOGY',
    RHEUMATOLOGY = 'RHEUMATOLOGY',
    UROLOGY = 'UROLOGY',
    OTHER = 'OTHER'
}